+++ experimental/packages/opentelemetry-instrumentation/src/autoLoader.ts
@@ -15,7 +15,7 @@
  */
 
 import { trace, metrics } from '@opentelemetry/api';
-import { logs } from '@opentelemetry/api-logs';
+import { logs } from '@opentelemetry/api/experimental';
 import {
   disableInstrumentations,
   enableInstrumentations,