+++ experimental/packages/sdk-logs/test/common/export/ConsoleLogRecordExporter.test.ts
@@ -16,7 +16,7 @@
 
 import * as assert from 'assert';
 import * as sinon from 'sinon';
-import { SeverityNumber } from '@opentelemetry/api-logs';
+import { SeverityNumber } from '@opentelemetry/api/experimental';
 
 import {
   LoggerProvider,